requirements.txt
ultralytics
roboflow
Pillow
pandas
openpyxl
requests
opencv-python
numpy
scipy
python-dotenv
pyyaml
fastapi
uvicorn
python-multipart
google-generativeai
